[didelphis-common-55] Solve the problem of replacing classes with transpilable alternatives
